Fix routes, add routes, and other changes
- Add backend/routes/middleware/showsMiddleware.js
  - has middleware functions from => backend/routes/shows.js
- Changes to backend/routes/shows.js
  - remove middleware functions from the above and use the above as
    imported modules
  - other small changes
- Changes to backend/routes/singleShow.js
  - add a provision for no user returned to the GET single show route
  - correct the POST new comment route so show_id is added via req.params
- Changes to backend/routes/singleUser.js
  - add route to GET all genres
  - add provision for no user returned to the GET single user route
- other small changes
