+++ backend/routes/shows.js
@@ -14,80 +14,21 @@
  * `/shows/user/:id`  : done
  */
 const express = require('express');
+const middleware = require('./middleware/showsMiddleware');
 
 const Show = require('../models/show');
-const User = require('../models/user');
-
-const { body, validationResult } = require('express-validator/check');
 
 const router = express.Router();
 
 /**
  * Middleware functions
- *
  */
-const getDistinctShowsAndUsers = async (req, res, next) => {
-  try {
-    const queryShows = await Show.findAll();
-
-    const reduced = queryShows.reduce((acc, show, i) => {
-      const { id, title, img_url, genre_id, user_id } = show.dataValues;
-      const showObj = {
-        ids: [],
-        users_watching: [],
-        title: title,
-        img_url: img_url,
-        genre_id: genre_id,
-        has_repeats: false
-      };
-
-      const showExists = acc.some((accShow, j) =>{
-        return accShow.title === title;
-      });
-
-      if (showExists === true && acc.length !== 0) {
-        const indexExisting = acc.findIndex((accShow) =>{
-            return accShow.title === title;
-        });
-        acc[indexExisting].users_watching.push(user_id);
-        acc[indexExisting].ids.push(id);
-        acc[indexExisting].has_repeats = true;
-        return acc;
-      } else {
-        showObj.users_watching.push(user_id);
-        showObj.ids.push(id);
-        acc.push(showObj);
-        return acc;
-      }
-      }, []);
-    res.locals.distinctShows = reduced;
-  } catch(error) {
-    res.status(500).send();
-  }
-  next();
-}
-const getUsers = async (req, res, next) => {
-  try {
-    const queryUsers = await User.findAll();
-    const users = queryUsers.map(user => {
-      return {
-        id: user.dataValues.id,
-        username: user.dataValues.username
-      }});
-    res.json({
-      shows: res.locals.distinctShows,
-      users: users
-    });
-  } catch(error) {
-    res.status(500).send();
-  }
-  next();
-}
+const { getAllShowsAndSort, getUsers } = middleware;
 
 /**
  * GET all DISTINCT shows
  */
-router.get('/', getDistinctShowsAndUsers, getUsers);
+router.get('/', getAllShowsAndSort, getUsers);
 
 /**
  * GET all shows for a user
@@ -98,9 +39,9 @@ router.get('/user/:id', async (req, res, next) => {
       user_id: req.params.id
     }
   });
-  if(queryUsers.length === 0) {
+  if(queryUsers == null || queryUsers.length === 0) {
     // TODO flesh this out more
-    res.status(409).send();
+    res.status(404).send();
   } else {
     const shows = queryShows.map(show => {
       const { id, title, img_url, genre_id } = show.dataValues
@@ -119,16 +60,15 @@ router.get('/user/:id', async (req, res, next) => {
  * GET all shows for a genre and for all users
  */
 router.get('/genre/:id', async (req, res, next) => {
-  // TODO: Add provisions for handling genre id not found
   //  attributes: [Sequelize.fn('DISTINCT', Sequelize.col('title')) ,'title'],
   const queryShows = await Show.findAll({
     where: {
       genre_id: req.params.id
     }
   });
-  if(queryUsers.length === 0) {
+  if(queryShows == null || queryShows.length === 0) {
     // TODO flesh this out more
-    res.status(409).send();
+    res.status(404).send();
   } else {
     const shows = queryShows.map(show => {
       const { id, title, img_url } = show.dataValues
