+++ backend/routes/singleShow.js
@@ -9,7 +9,7 @@
  *
  * Status:
  * `/show/:id` : GET requests NOT complete
- * `/show/:id` : POST request partially comlete
+ * `/show/:id` : POST request comlete
  */
 const express = require('express');
 const Show = require('../models/show');
@@ -19,6 +19,7 @@ const { body, validationResult } = require('express-validator/check');
 
 const router = express.Router();
 
+// TODO: split this into two middleware functions and unify it later
 /**
  * GET single show for a single user
  */
@@ -29,10 +30,15 @@ router.get('/:id', async (req, res) => {
         id: req.params.id
       }
     });
-    // TODO: Add thing in to return something if no user can be found
-    const singleShow = queryShows.dataValues;
-    res.json({show: singleShow});
+    if(queryShows == null || queryShows.length === 0) {
+      // TODO flesh this out more
+      res.status(404).send();
+    } else {
+      const singleShow = queryShows.dataValues;
+      res.json({show: singleShow});
+    }
   } catch(error) {
+    console.log(error);
     res.status(500).send();
   }
 });
@@ -64,12 +70,10 @@ router.post('/:id', [
 ], async (req, res) => {
   /**
    * NOTE: user_id and show_id should be surmised from hidden form inputs
-   * TODO: add validation
    */
   let errors = validationResult(req);
-  const { user_id, show_id, comment_body  } = req.body;
+  const { user_id, comment_body  } = req.body;
   try {
-    // TODO: Add thing in to return something if no user can be found
     if (!errors.isEmpty()) {
       res.status(400).json({
         errors: errors.mapped()
@@ -77,7 +81,7 @@ router.post('/:id', [
     } else {
       await Comment.create({
         user_id: user_id,
-        show_id: show_id,
+        show_id: req.params.id,
         comment_body: comment_body
       });
       res.status(200).json({msg: "Success, comment added"}).send();
