+++ backend/routes/users.js
@@ -12,7 +12,6 @@
  */
 const express = require('express');
 const User = require('../models/user');
-const Genre = require('../models/genre');
 
 const router = express.Router();
